feat(activity): add activity screen, set up training selection params

+++ lib/domain/models/training.dart
@@ -10,8 +10,8 @@ abstract class Training with _$Training {
   const factory Training({
     @JsonKey(name: 'training_id') required int id,
     @JsonKey(name: 'training_name') required String name,
-    @JsonKey(name: 'attack_points') @Default(1) int attackPoints,
-    @JsonKey(name: 'deffence_points') @Default(1) int defensePoints,
+    @JsonKey(name: 'attack_points') @Default(1.0) double attackPoints,
+    @JsonKey(name: 'deffence_points') @Default(1.0) double defensePoints,
   }) = _Training;
 
   factory Training.fromJson(Map<String, dynamic> json) => _$TrainingFromJson(json);
